+++ plc4go/internal/plc4go/bacnetip/readwrite/model/BACnetPropertyStates.go
@@ -94,7 +94,7 @@ func (m *BACnetPropertyStates) GetClosingTag() *BACnetClosingTag {
 /////////////////////// Accessors for virtual fields.
 ///////////////////////
 func (m *BACnetPropertyStates) GetPeekedTagNumber() uint8 {
-	return m.GetPeekedTagHeader().GetActualTagNumber()
+	return uint8(m.GetPeekedTagHeader().GetActualTagNumber())
 }
 
 ///////////////////////
