+++ plc4go/internal/plc4go/bacnetip/readwrite/model/BACnetTagPayloadCharacterString.go
@@ -72,7 +72,7 @@ func (m *BACnetTagPayloadCharacterString) GetValue() string {
 /////////////////////// Accessors for virtual fields.
 ///////////////////////
 func (m *BACnetTagPayloadCharacterString) GetActualLengthInBit() uint16 {
-	return uint16(uint16(m.ActualLength)*uint16(uint16(8))) - uint16(uint16(8))
+	return uint16(uint16(uint16(m.ActualLength)*uint16(uint16(8))) - uint16(uint16(8)))
 }
 
 ///////////////////////
